+++ Qualys_Virtual_Scanner_Appliance/Azure_Terraform_Template/changelog.md
@@ -1,4 +1,8 @@
 # Azure QVSA Deployment
+
+## Version: v1.2.0
+Code updation for deleting userdata.
+
 ## Version: v1.1.0
 This version introduces improvements to streamline the deployment of Scanner Appliances using existing resources across Azure.