+++ Qualys_Virtual_Scanner_Appliance/GCP_Terraform_Template/changelog.md
@@ -1,4 +1,8 @@
 # GCP QVSA Deployment
+
+## Version: v1.2.0
+Support for Google Shielded VM features
+
 ## Version: v1.1.0
 This version introduces improvements to streamline the deployment of Scanner Appliances using existing resources across GCP.